requirements.txt
pandas
openpyxl
pypdf
python-docx
python-pptx
xlrd
pyxlsb
